hat-or-banner

+++ Content/MissileAddons/BeamCombos/PlasmaMachinegun.cs
@@ -180,7 +180,7 @@ namespace MetroidMod.Content.MissileAddons.BeamCombos
 					Color color23 = color2;
 					color23 = P.GetAlpha(color23);
 					color23 *= (amt - i) / ((float)amt);
-					color23.A = (byte)((float)color23.A * ((float)(amt - i) / (float)amt));
+					color23.A = (byte)(color23.A * ((amt - i) / (float)amt));
 					float scale = MathHelper.Lerp(P.scale, P.scale * scaleDrop, (float)i / amt);
 
 					float vel2 = Math.Min(Vector2.Distance(P.oldPos[i] + (P.Size / 2f), startPos), P.velocity.Length());
@@ -190,7 +190,7 @@ namespace MetroidMod.Content.MissileAddons.BeamCombos
 						{
 							Color color4 = color23;
 							color4 *= (float)(vel2 - j) / ((float)vel2);
-							color4.A = (byte)((float)color4.A * ((float)(vel2 - j) / (float)vel2));
+							color4.A = (byte)(color4.A * ((float)(vel2 - j) / (float)vel2));
 							Vector2 oldPos = P.oldPos[i] + (P.Size / 2f) - (Vector2.Normalize(P.velocity) * j);
 							sb.Draw(tex, oldPos - Main.screenPosition, new Rectangle?(new Rectangle(0, y4, tex.Width, height)),
 							color23, P.oldRot[i], new Vector2(tex.Width / 2f, P.height / 2f), scale, effects, 0f);
@@ -207,7 +207,7 @@ namespace MetroidMod.Content.MissileAddons.BeamCombos
 				{
 					Color color3 = P.GetAlpha(color2);
 					color3 *= (float)(vel - j) / ((float)vel);
-					color3.A = (byte)((float)color3.A * ((float)(vel - j) / (float)vel));
+					color3.A = (byte)(color3.A * ((float)(vel - j) / (float)vel));
 					Vector2 pos = P.Center - (Vector2.Normalize(P.velocity) * j);
 					sb.Draw(tex, pos - Main.screenPosition, new Rectangle?(new Rectangle(0, y4, tex.Width, height)),
 					P.GetAlpha(color2), P.rotation, new Vector2(tex.Width / 2f, P.height / 2f), P.scale, effects, 0f);
